refactor(mobile): move camera modal to shared components
