+++ apps/mobile/src/shared/components/CameraModal/useCamera.ts
